web/requirements.txt
boto3==1.23.9
botocore==1.26.10
click==8.1.3
Flask==2.1.2
Flask-Login==0.6.1
Flask-WTF==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.9
Werkzeug==2.1.2
WTForms==3.0.1
